% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_activity.R
\name{fit_activity}
\alias{fit_activity}
\title{Fit activity data}
\usage{
fit_activity(data, wells_output, normalise = FALSE)
}
\arguments{
\item{data}{The TRA data (dataframe).}

\item{wells_output}{File path for the per-well output figure (character).}

\item{normalise}{Wether to normalise the data or not (boolean).}
}
\value{
A dataframe containing the calculated activities.
}
\description{
Fit activity data to a sigmoidal curve.
}
\details{
This are functions for fitting a sigmoidal curve to each TRA well
using the time-series data therein, yielding the activitity for a given
endolysine and concentration.
}
\examples{
# Get input file path
file_path <- system.file("extdata", "example_data_1.csv",
                         package = "ATRADA")

# Read example data
data <- read_tra_data(file_path)

# Get the activity for each well
activity <- fit_activity(data, "wells.png")
}
