% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tra.R
\name{read_tra_data}
\alias{read_tra_data}
\title{Read TRA data}
\usage{
read_tra_data(file, skip_rows = 0)
}
\arguments{
\item{file}{The input CSV file (character).}

\item{skip_rows}{The number of rows to skip (integer).}
}
\value{
A dataframe with the read TRA data.
}
\description{
Read TRA data from CSV files.
}
\details{
This function reads TRA data from CSV files for use in downstream analyses.
}
\examples{
# Get input file path
file_path <- system.file("extdata", "example_data_1.csv",
                         package = "ATRADA")

# Read TRA data from file
data <- read_tra_data(file_path)
data <- read_tra_data(file_path, skip_rows = 5)
}
